backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
telethon==1.36.0
python-multipart==0.0.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
PyJWT[crypto]==2.8.0
aiofiles==23.2.1
Pillow
websockets==12.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0